requirements.txt
spotipy
RPi.GPIO
mfrc522
